+++ lib/src/dialogs/x_dialog.dart
@@ -4,14 +4,18 @@ import "package:flutter/material.dart";
 
 import "../../x_containers.dart";
 
-/// A custom [AlertDialog]
+/// A custom [AlertDialog] matching the style of XContainers.
+///
+/// It features a title and a content widget and two callback buttons.
+/// The buttons are traditionally used for "yes" and "no" behaviors
+/// but can be renamed and will not be displayed if their name is set to [null].
 class XDialog {
   // VARIABLES =================================================================
 
-  /// The text to display at the top of the dialog.
+  /// {@macro x_containers.docs.title}
   final Widget title;
 
-  /// The description of the effect of the dialog.
+  /// {@macro x_containers.docs.content}
   final Widget? content;
 
   /// The text displayed on the cancel button.
@@ -24,7 +28,7 @@ class XDialog {
   /// If it is [null] then the button isn't displayed.
   final String? validateText;
 
-  /// The color of the background of the box.
+  /// {@macro x_containers.docs.color}
   final Color? color;
 
   /// The behavior to execute when the "cancel" button is pressed.
@@ -44,6 +48,12 @@ class XDialog {
   // CONSTRUCTOR ===============================================================
 
   /// Returns an instance of [XDialog] matching the given parameters.
+  ///
+  /// - [title] and [content] are widgets.
+  /// - [cancelText] and [validateText] are the Strings displayed on the buttons.
+  /// - [onCancel] and [onValidate] are the callbacks run when the buttons are pressed.
+  /// - [color] is the background color of the dialog.
+  /// - [backgroundBlur] is the intensity of the gaussian blur applied to stuff that is "behind" the dialog.
   const XDialog({
     required this.title,
     this.content,
@@ -56,6 +66,12 @@ class XDialog {
   });
 
   /// Returns an XDialog displaying only text messages.
+  ///
+  /// - [title] and [content] are Strings.
+  /// - [cancelText] and [validateText] are the Strings displayed on the buttons.
+  /// - [onCancel] and [onValidate] are the callbacks run when the buttons are pressed.
+  /// - [color] is the background color of the dialog.
+  /// - [backgroundBlur] is the intensity of the gaussian blur applied to stuff that is "behind" the dialog.
   factory XDialog.text({
     required String title,
     String? content,
@@ -63,6 +79,7 @@ class XDialog {
     String? validateText = "Okay",
     void Function()? onCancel,
     void Function()? onValidate,
+    Color? color,
     double backgroundBlur = 1,
   }) =>
       XDialog(
@@ -72,6 +89,7 @@ class XDialog {
         validateText: validateText,
         onCancel: onCancel,
         onValidate: onValidate,
+        color: color,
         backgroundBlur: backgroundBlur,
       );
 
@@ -87,31 +105,40 @@ class XDialog {
           sigmaY: backgroundBlur,
         ),
         child: AlertDialog(
+          contentPadding: xTheme.padding,
           backgroundColor: color ?? Theme.of(context).cardColor,
           shape:
               RoundedRectangleBorder(borderRadius: xTheme.dialogBorderRadius),
           title: DefaultTextStyle(
-              style:
-                  Theme.of(context).textTheme.titleSmall ?? const TextStyle(),
-              child: title),
+            style: const TextStyle()
+                .merge(Theme.of(context).textTheme.titleMedium),
+            child: title,
+          ),
           content: DefaultTextStyle(
-              style:
-                  Theme.of(context).textTheme.bodyMedium ?? const TextStyle(),
-              child: content ?? const SizedBox()),
+            style:
+                const TextStyle().merge(Theme.of(context).textTheme.bodyMedium),
+            child: content ?? const SizedBox(),
+          ),
           actions: <Widget>[
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-                onCancel?.call();
-              },
-              child: Text(cancelText ?? "Cancel"),
+            Visibility(
+              visible: cancelText != null,
+              child: TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop();
+                  onCancel?.call();
+                },
+                child: Text(cancelText ?? "Cancel"),
+              ),
             ),
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-                onValidate?.call();
-              },
-              child: Text(validateText ?? "Okay"),
+            Visibility(
+              visible: validateText != null,
+              child: TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop();
+                  onValidate?.call();
+                },
+                child: Text(validateText ?? "Okay"),
+              ),
             ),
           ],
         ),