+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## 1.2.0
+
+- `ADDED`: [XListTile] as a way to have a box-less reusable tile layout.
+- `ADDED`: A preset [TextTheme] to [xTheme.getTheme] that can be fully or partially overridden.
+- `CHANGED`: [XCard] and [XSnackbar] use [XListTile].
+- `CHANGED`: The documentation which is now handled in part by macros.
+- `CHANGED`: The example to showcase changing the text theme through [xTheme].
+
+- `BREAKING`: [XSnackbar] doesn't use the [titleStyle] and [contentStyle] properties which are handled by the text theme.
+
+## 1.1.1
+
+- `CHANGED`: Fixed library name (from `x_container` to `x_containers`).
+
+## 1.1.0
+
+- `ADDED`: a makefile containing a few utility commands.
+- `IMPROVED`: The documentation comments on the class constructors.
+- `IMPROVED`: The [XDialog] buttons turn invisible if their label is set to [null].
+
+- `BREAKING`: In [XSnackbar], the field [message] is renamed to [content] to match the naming of other classes.
+- `BREAKING`: In [XCard], the field [subtitle] is renamed to [content] to match the naming of other classes.
+
+- `CHANGED`: The example to match the breaking changes.
+
+## 1.0.3-dev.1
+
+- `IMPROVED`: The README for even better legibility.
+- `IMPROVED`: Using asset image from github.
+
 ## 1.0.2
 
 - `ADDED`: An XContainers preview image for the README.
@@ -63,7 +93,6 @@
 - `ADDED`: XCard.
 - `ADDED`: XTheme to manage the shared properties.
 
----
 
 ## How to use
 
@@ -78,6 +107,7 @@ Keywords:
 - `CHANGED`: When you switch some code for different code (for instance, if we decided to radically change the way the movies are displayed).
 - `DELETED`: When you remove some feature or code.
 - `REFACTORED`: When you change file names or a linting rule.
+- `BREAKING`: Any breaking change.
 
 ## WARNING